[FIX] base: fix file extension when using default name
This commit aims to fix an issue when downloading attachments and using
'name' as the field for the file's name.
When 'name' contains a dot, the file is considered to already have a
file extension, and thus the right extension is not added to the name.
After this commit, using 'name' as a filename field will also consider
the name to be a 'default' for any other model than `ir.attachment`.

TaskId-2826061

closes odoo/odoo#90789

X-original-commit: ca854b66f9bfcfa0c21b96ce82710e2d8f4bd8c6
Signed-off-by: Julien Castiaux <juc@odoo.com>
Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>